+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 [unreleased]
 
 - Better query performance for `timestamptz` columns
+- Added support for schemas other than `public`
 
 ## 0.2.3